S3StorageService: implemented source file deletion

+++ backend/src/main/java/de/htwg/klaut/backend/service/Word2VecModelService.java
@@ -168,7 +168,15 @@ public class Word2VecModelService implements IModelService<Word2VecParams> {
     public void deleteModel(CompositeId modelId) throws ModelNotFoundException, SourceNotFoundException {
         log.debug("deleting model {}", modelId);
 
-        s3StorageService.deleteFilesForId(modelId);
+        Optional<Model> modelOptional = modelRepository.findById(modelId);
+        if (!modelOptional.isPresent()) {
+            throw new ModelNotFoundException(modelId);
+        }
+
+        final Model modelToDelete = modelOptional.get();
+        for (String sourceUrlToDelete : modelToDelete.getSourceUrls()) {
+            s3StorageService.deleteSourceFile(sourceUrlToDelete);
+        }
         modelRepository.deleteById(modelId);
     }
 }
